// src/types/contentful.d.ts
export interface ILinkFields {
  /** Internal Name */
  name?: string | undefined;

  /** Label */
  text: string;

  /** Page */
  page?: IPage | IPageSimple | IPost | undefined;

  /** URL */
  url?: string | undefined;

  /** New Window */
  newWindow?: boolean | undefined;
}

/** Call to Action */

export interface ILink extends Entry<ILinkFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "link";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}
// ...
export interface INavItem extends Entry<INavItemFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "navItem";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}
// ...
export interface IPage extends Entry<IPageFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "page";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface IPageSimpleFields {
  /** Title */
  title: string;

  /** Slug */
  slug: string;

  /** Text */
  text?: Document | undefined;

  /** Metadata */
  metadata?: ISeo | undefined;

  /** Display this translation? */
  translated?: boolean | undefined;
}

export interface IPageSimple extends Entry<IPageSimpleFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "pageSimple";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}
// ...
export interface IPost extends Entry<IPostFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "post";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}
// ...
export interface ISeo extends Entry<ISeoFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "seo";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface ISettingsFields {
  /** Site Title */
  title: string;

  /** Legal Name */
  legalName?: string | undefined;

  /** Metadata */
  metadata?: ISeo | undefined;

  /** Logo Svg */
  logoSvg?: string | undefined;

  /** Logo Image */
  logoImage?: Asset | undefined;

  /** Image Quality */
  imagesQuality?: number | undefined;

  /** Main Navigation */
  mainNavigation?: (ILink | ISimpleList | INavItem)[] | undefined;

  /** Nav CTA 1 */
  navCta1?: ILink | undefined;

  /** Nav CTA 2 */
  navCta2?: ILink | undefined;

  /** Footer Navigation */
  footerNavigation?: (ILink | ISimpleList)[] | undefined;

  /** Legal Navigation */
  legalNavigation?: ILink[] | undefined;

  /** Facebook */
  facebook?: string | undefined;

  /** Twitter */
  twitter?: string | undefined;

  /** Instagram */
  instagram?: string | undefined;

  /** LinkedIn */
  linkedIn?: string | undefined;

  /** Pinterest */
  pinterest?: string | undefined;

  /** Google Analytics ID */
  googleAnalyticsId?: string | undefined;

  /** Google Tag Manager ID */
  googleTagManagerId?: string | undefined;

  /** CookieBot Id */
  cookieBotId?: string | undefined;

  /** Facebook App ID */
  facebookAppId?: string | undefined;

  /** Greenhouse Board ID */
  greenhouseBoardToken?: string | undefined;

  /** Marker.io Destination ID */
  markerIoDestinationId?: string | undefined;

  /** Marketo Hostname */
  marketoHostname?: string | undefined;

  /** Marketo Munchkin ID */
  marketoMunchkinId?: string | undefined;

  /** Enable Search */
  enableSearch?: boolean | undefined;
}

export interface ISettings extends Entry<ISettingsFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "settings";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface ISimpleListFields {
  /** Internal Name */
  name?: string | undefined;

  /** Title */
  title?: string | undefined;

  /** Items */
  items?: ILink[] | undefined;
}

/** Specific to navigation. Used for sub-navigation list */

export interface ISimpleList extends Entry<ISimpleListFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "simpleList";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}